\name{acc.roxygen2-package}
\alias{acc.roxygen2-package}
\alias{acc.roxygen2}
\docType{package}
\title{
Fuctions to Extend roxygen2
}
\description{
A (eventual) collection of tools to make roxygen2 even easier to use (note these functions are designed for windows users and may be extended to Mac)
}
\details{
\tabular{ll}{
Package: \tab acc.roxygen2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-06\cr
License: \tab GPL-2\cr
}
}
\author{
Tyler Rinker

Maintainer: Tyler Rinker <tyler.rinker@gmail.com>
}