\name{doc_check}
\alias{doc_check}
\title{Generate Text of Roxygen Documentation}
\usage{
  doc_check(in_path = "C:/Users/trinker/GitHub/qdap/R",
    out_path = paste0(dt, "spell_check.doc"))
}
\arguments{
  \item{in_path}{Path to the R documentation in a repo.}

  \item{out_path}{The desired file out path.}
}
\description{
  Generates text of roxygen documentation for quick spell
  check.
}

