\name{examples}
\alias{examples}
\title{Generate File of Package Examples}
\usage{
  examples(file = "examples.txt",
    path = "C:/Users/trinker/GitHub/qdap/R/")
}
\arguments{
  \item{file}{A file name, however, not a path.}

  \item{path}{Path to repo.}
}
\description{
  Generates a file of examples to your working directory.
}

