% Generated by roxygen2 (4.0.2): do not edit by hand
\name{examples}
\alias{examples}
\title{Generate File of Package Examples}
\usage{
examples(repo = getOption("primary_repo"), file = file.path(getwd(),
  "examples.txt"), base.git = getOption("base_git"))
}
\arguments{
\item{repo}{Repo name.}

\item{file}{A character string naming the file to print to.}

\item{base.git}{Base path to where repos are store (the repo is stored).}
}
\description{
Generates a file of examples to your working directory.
}

