% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand_statdoc}
\alias{expand_statdoc}
\title{Expand staticdocs index.html File}
\usage{
expand_statdoc(path, file = NULL, to.icon = NULL, readme = NULL,
  combine = NULL, drop = NULL, rm.other = TRUE)
}
\arguments{
\item{path}{Path to the index.html file.}

\item{file}{The path/file name to output.  If NULL the original index is
overwritten.}

\item{to.icon}{The extra functions found in the parenthesis for who new icons
should be added.}

\item{readme}{Path to the readme file.}

\item{combine}{A list of vectors to combine together.  The functions will be
combined int he order of each vector.}

\item{drop}{A character vector of functions in the parenthesis that should be
excluded from the index file.}
}
\description{
staticdocs writes multiple files contained in an .Rd file to one icon
followed by a parenthesis with the other functions.  This function exapands
the parenthesis to be separate icons of their own.
}
\examples{
\dontrun{
(examp <- system.file("extdata/qdap", package = "acc.roxygen2"))
file.copy(examp, to=getwd(), recursive = TRUE)
path <- paste0(getwd(), "/qdap/index.html")
file <- paste0(getwd(), "/qdap/index2.html")

library(qdap)
expand_statdoc(path, file, drop = qcv(syn, mgsub, adjmat, wc))

file2 <- paste0(getwd(), "/qdap/index3.html")
extras <- qcv(right.just, coleman_liau, flesch_kincaid, fry, linsear_write, SMOG)
expand_statdoc(path, file2, to.icon = extras, drop = qcv(syn, mgsub, adjmat, wc))

file3 <- paste0(getwd(), "/qdap/index4.html")
rdme <- system.file("extdata/readme.R", package = "acc.roxygen2")
expand_statdoc(path, file3, to.icon = extras,
    readme = rdme, combine = qcv(character.table, char.table),
    drop = qcv(syn, mgsub, adjmat, wc))
}
}

