% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search_repo}
\alias{search_repo}
\title{Search a Repo for Terms}
\usage{
search_repo(..., repo = getOption("primary_repo"),
  base.git = getOption("base_git"), terms = NULL, split = "\\\\|",
  ignore.case = FALSE, perl = FALSE, fixed = FALSE)
}
\arguments{
\item{repo}{Repo name.}

\item{base.git}{Base path to where repos are store (the repo is stored).}

\item{terms}{An optional argument to present the terms as one long character
string.  This is useful if the split (separator) is not a comma (e.g. spaces
are the term separators).}

\item{split}{Character vector of length one to use for splitting (i.e. the
separator used in the vector).  For use with the argument \code{terms}.}

\item{ignore.case}{logical.  If \code{FALSE}, the pattern matching is case
sensitive and if \code{TRUE}, case is ignored during matching.}

\item{perl}{logical. Should Perl-compatible regexps be used?}

\item{fixed}{logical. If \code{TRUE}, pattern is a string to be matched as
is. Overrides all conflicting arguments.}

\item{\ldots}{Terms to search for in the .R files.  No need for quotes.}
}
\description{
Generates counts of terms found in the .R documentation.
}

